Consolidate the location model and schema code since there are several dependencies between them.
